# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';
import type { SalesPoint } from './salesPoint';

export interface DashboardSummary {
  revenue: number;
  revenueDelta: number;
  pendingOrders: number;
  totalOrders: number;
  productsSold: number;
  activeProducts: number;
  recentOrders: Order[];
  salesByDay: SalesPoint[];
}