# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDeliveryType } from './orderDeliveryType';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName: string;
  /** @nullable */
  username: string | null;
  telegramId: string;
  productName: string;
  amount: number;
  currency: string;
  status: OrderStatus;
  deliveryType: OrderDeliveryType;
  createdAt: Date;
  /** @nullable */
  deliveredAt: string | null;
  canRedeliver: boolean;
}