# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductDeliveryType } from './productDeliveryType';

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  currency: string;
  deliveryType: ProductDeliveryType;
  deliveryLabel: string;
  inventory: number;
  active: boolean;
  createdAt: Date;
}