# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputDeliveryType } from './productInputDeliveryType';

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  currency: string;
  deliveryType: ProductInputDeliveryType;
  /** @minLength 1 */
  deliveryLabel: string;
  /** @minimum 0 */
  inventory: number;
}