# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductUpdateDeliveryType } from './productUpdateDeliveryType';

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  currency?: string;
  deliveryType?: ProductUpdateDeliveryType;
  /** @minLength 1 */
  deliveryLabel?: string;
  /** @minimum 0 */
  inventory?: number;
  active?: boolean;
}