# lib/api-zod/src/generated/types/supportTicket.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SupportTicketStatus } from './supportTicketStatus';

export interface SupportTicket {
  id: number;
  customerName: string;
  /** @nullable */
  username: string | null;
  topic: string;
  status: SupportTicketStatus;
  lastMessage: string;
  updatedAt: Date;
}